requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.0
sqlalchemy~=2.0.40
pydantic[email]~=2.11.2
passlib[bcrypt]
python-jose~=3.4.0
python-multipart
python-dotenv~=1.1.0
pymysql
starlette
bcrypt
fastapi-pagination